requirements.txt
beautifulsoup4==4.9.1
dash-bio==0.5.0
dash-bootstrap-components==0.10.7
dash-daq==0.5.0
helper==2.4.2
imagehash==4.2.0
jupyterlab==2.2.9
pandas-profiling==2.9.0
pip-chill==1.0.0
psutil==5.7.3
pyarrow==2.0.0
ruamel.yaml==0.16.12
sqlalchemy==1.3.20
statsmodels==0.12.1
